+++ programs/bpf/Cargo.toml
@@ -26,7 +26,7 @@ solana-bpf-loader-api = { path = "../bpf_loader_api", version = "0.19.0-pre0" }
 solana-logger = { path = "../../logger", version = "0.19.0-pre0" }
 solana-runtime = { path = "../../runtime", version = "0.19.0-pre0" }
 solana-sdk = { path = "../../sdk", version = "0.19.0-pre0" }
-solana_rbpf = "=0.1.15"
+solana_rbpf = { path = "../../../rbpf" } # "=0.1.15"
 
 [[bench]]
 name = "bpf_loader"
