+++ programs/bpf/rust/sysval/Cargo.toml
@@ -2,9 +2,9 @@
 # Note: This crate must be built using build.sh
 
 [package]
-name = "solana-bpf-rust-clock"
+name = "solana-bpf-rust-sysval"
 version = "0.19.0-pre0"
-description = "Solana BPF clock sysvar test"
+description = "Solana BPF sysvar test"
 authors = ["Solana Maintainers <maintainers@solana.com>"]
 repository = "https://github.com/solana-labs/solana"
 license = "Apache-2.0"
@@ -23,4 +23,4 @@ members = []
 
 [lib]
 crate-type = ["cdylib"]
-name = "solana_bpf_rust_clock"
+name = "solana_bpf_rust_sysval"
