+++ programs/bpf_loader_api/Cargo.toml
@@ -16,7 +16,7 @@ log = "0.4.8"
 serde = "1.0.100"
 solana-logger = { path = "../../logger", version = "0.19.0-pre0" }
 solana-sdk = { path = "../../sdk", version = "0.19.0-pre0" }
-solana_rbpf = "=0.1.15"
+solana_rbpf = { path = "../../../rbpf" } # "=0.1.15"
 
 [lib]
 crate-type = ["lib"]
